Moved some data that was in helpers into 'data' directory
