+++ src/data/directory-regex.ts
