+++ src/data/globals.ts
