+++ src/helpers/index.ts
@@ -1,12 +1,9 @@
-export { default as globals } from './globals';
 export { default as getConfigFile } from './get-config-file';
-export { default as directoryRegex } from './directory-regex';
 export { default as skip } from './skip';
 export { default as printMessage } from './print-message';
 export { default as printLogo } from './print-logo';
 export { default as nameError } from './throw-error';
 export { default as askQuestion } from './ask-question';
-export { default as questions } from './questions';
 export { default as throwError } from './throw-error';
 export { default as randomWordGenerator } from './random-word-generator';
 export { default as getFileListFromRegistry } from './get-file-list-from-registry';
