+++ src/helpers/print-logo.ts
@@ -1,5 +1,5 @@
 import chalk from 'chalk';
-import globals from './globals';
+import globals from '@data/globals';
 
 export default () => {
   return console.log(
